requirements.txt
Flask
requests
python-dotenv
PyPDF2
Werkzeug
